feat(stories): isolate storybook
