--- .storybook/babel.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    presets: [
-        ['@babel/preset-env', { loose: true }],
-        ['@babel/preset-react', { runtime: 'automatic' }],
-        '@babel/preset-typescript',
-    ],
-}
