--- .storybook/manager.js
@@ -1,8 +1,0 @@
-import { create } from '@storybook/theming'
-
-export default create({
-    base: 'light',
-    brandTitle: 'nivo',
-    brandUrl: 'https://nivo.rocks',
-    brandImage: null,
-})
