--- .storybook/preview-head.html
@@ -1,26 +1,0 @@
-<style type="text/css">
-#root {
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol';
-    text-rendering: optimizeLegibility;
-    -webkit-font-smoothing: antialiased;
-}
-
-#story-root {
-    /*
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    */
-    font-family: 'Consolas', monospace;
-    font-size: 13px;
-}
-
-#story-root text {
-    font-family: 'Consolas', monospace;
-}
-</style>
