+++ packages/core/index.d.ts
@@ -36,11 +36,7 @@ export type BoxAlign =
     | 'bottom-left'
     | 'left'
 export const boxAlignments: BoxAlign[]
-export function alignBox(
-    box: AlignBox,
-    container: AlignBox,
-    alignment: BoxAlign
-): [number, number]
+export function alignBox(box: AlignBox, container: AlignBox, alignment: BoxAlign): [number, number]
 
 export type GetColor<T> = (datum: T) => string
 export type Colors = string[] | string
@@ -370,11 +366,7 @@ export function degreesToRadians(degrees: number): number
 export function radiansToDegrees(radians: number): number
 export function absoluteAngleDegrees(degrees: number): number
 export function normalizeAngle(degrees: number): number
-export function clampArc(
-    startAngle: number,
-    endAngle: number,
-    length?: number
-): [number, number]
+export function clampArc(startAngle: number, endAngle: number, length?: number): [number, number]
 
 type Accessor<T extends keyof U, U> = T extends string ? U[T] : never
 
@@ -450,9 +442,7 @@ export function positionFromAngle(
 export type ValueFormat<Value, Context = void> =
     | string // d3 formatter
     // explicit formatting function
-    | (Context extends void
-          ? (value: Value) => string
-          : (value: Value, context: Context) => string)
+    | (Context extends void ? (value: Value) => string : (value: Value, context: Context) => string)
 export function getValueFormatter<Value, Context = void>(
     format?: ValueFormat<Value, Context>
 ): Context extends void ? (value: Value) => string : (value: Value, context: Context) => string
@@ -501,10 +491,7 @@ interface CartesianMarkersProps<
     yScale: (value: Y) => number
     markers: CartesianMarkerProps<X | Y>[]
 }
-type CartesianMarkersType = <
-    X extends DatumValue = DatumValue,
-    Y extends DatumValue = DatumValue
->(
+type CartesianMarkersType = <X extends DatumValue = DatumValue, Y extends DatumValue = DatumValue>(
     props: CartesianMarkersProps<X, Y>
 ) => JSX.Element
 export const CartesianMarkers: CartesianMarkersType
