+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
   - 'packages/*'
   - 'api/'
   - 'website/'
+  - 'storybook/'
