+++ storybook/.storybook/preview.js
@@ -1,0 +1,14 @@
+/** @type { import('@storybook/react').Preview } */
+const preview = {
+    parameters: {
+        actions: { argTypesRegex: '^on[A-Z].*' },
+        controls: {
+            matchers: {
+                color: /(background|color)$/i,
+                date: /Date$/,
+            },
+        },
+    },
+}
+
+export default preview
