+++ storybook/stories/bar/ResponsivBarCanvas.stories._tsx
